Refactor Factory and SummaryReader.
We now have the ability to customize objects generated for every
supported language and/or blockchain.
