+++ lib/domain/libBlockchain/Factory.h
@@ -19,7 +19,9 @@ namespace vanguard {
     template<typename Base, typename Domain>
     class Blockchain<Base, Domain>::Factory : public Base::Factory {
     public:
-        typename Domain::Function *createBlkFn(const llvm::Module &module, BlockchainModel<Domain> &model, std::string contract, std::string name, std::string selector, bool isCnstr, Visibility vis, Mutability mut, std::vector<vanguard::Variable<Domain> *> params, std::vector<vanguard::Variable<Domain> *> rets, std::vector<std::string> mods) {
+        typename Domain::Function *createBlkFn(const llvm::Module &module, BlockchainModel<Domain> &model, std::string contract,
+                                               std::string name, std::string selector, bool isCnstr, Visibility vis, Mutability mut,
+                                               std::vector<vanguard::Variable<Domain> *> params, std::vector<vanguard::Variable<Domain> *> rets, std::vector<std::string> mods) {
             auto fn = new typename Domain::Function(name, selector, isCnstr, vis, mut, std::move(params), std::move(rets), std::move(mods), nullptr);
 
             for(auto &llvmFn : module) {
@@ -38,41 +40,36 @@ namespace vanguard {
             return contract;
         }
 
-        vanguard::StructType<Domain> *createStruct(const llvm::Module &mod, BlockchainModel<Domain> &model, const std::string& name, std::vector<vanguard::Variable<Domain> *> fields) {
-            auto *structType = new typename Domain::template BlkStructType<llvm::StructType>>(std::move(fields), nullptr);
+        StructType<Domain> *createStruct(const llvm::Module &mod, BlockchainModel<Domain> &model, const std::string& name, std::vector<vanguard::Variable<Domain> *> fields) {
+            auto *structType = new typename Domain::template BlkStructType<llvm::StructType>(std::move(fields), nullptr);
             return structType;
         }
 
-        vanguard::Variable<Domain> *createVariable(const llvm::Module &module, BlockchainModel<Domain> &model, std::string name, typename Domain::Type *type) {
+        Variable<Domain> *createVariable(const llvm::Module &module, BlockchainModel<Domain> &model, std::string name, typename Domain::Type *type) {
             auto *var = new typename Domain::template BlkVariable<llvm::Value>(std::move(name), type, nullptr);
             return var;
         }
 
-        vanguard::ArrayType<Domain> *createArrayType(const llvm::Module &module, BlockchainModel<Domain> &model, std::string name, typename Domain::Type *base) {
+        ArrayType<Domain> *createArrayType(const llvm::Module &module, BlockchainModel<Domain> &model, std::string name, typename Domain::Type *base) {
             auto *arr = new typename Domain::template BlkArrayType<llvm::ArrayType>(base, nullptr);
             return arr;
         }
 
-        vanguard::MapType<Domain> *createMapType(const llvm::Module &module, BlockchainModel<Domain> &model, std::string name, typename Domain::Type *keyType, typename Domain::Type *valType) {
+        MapType<Domain> *createMapType(const llvm::Module &module, BlockchainModel<Domain> &model, std::string name, typename Domain::Type *keyType, typename Domain::Type *valType) {
             auto *map = new typename Domain::template BlkMapType<llvm::Type>(keyType, valType, nullptr);
             return map;
         }
 
-        typename Domain::CompilationUnit *createBlkUnit(const llvm::Module *module, const std::string& summary) {
+        typename Domain::CompilationUnit *createBlkUnit(const llvm::Module *module, const std::vector<typename Domain::Contract *> &contracts) {
             if(module == nullptr) {
                 return nullptr;
             }
 
             if(this->moduleMap.find(module) == this->moduleMap.end()){
-                SummaryReader<Domain> reader(*module, summary);
-                this->moduleMap[module] = new typename Domain::CompilationUnit(reader.contracts(), module);
+                this->moduleMap[module] = new typename Domain::CompilationUnit(contracts, module);
             }
 
             return this->moduleMap[module];
-        }
-
-        typename Domain::Type *createBasicType(const llvm::Module &module, BlockchainModel<Domain> &model, std::string name) {
-            return nullptr;
         }
 
     protected:
