+++ lib/domain/libBlockchain/langs/CMakeLists.txt
@@ -1,0 +1,6 @@
+target_sources(libVanguard PRIVATE
+        LanguageFactory.cpp
+)
+
+add_subdirectory(rust)
+add_subdirectory(solidity)
