+++ lib/domain/libBlockchain/langs/rust/near/CMakeLists.txt
@@ -1,0 +1,3 @@
+target_sources(libVanguard PRIVATE
+  NearFactory.cpp
+)
