+++ lib/domain/libBlockchain/langs/rust/near/NearFactory.h
@@ -1,0 +1,29 @@
+//
+// Created by Kostas Ferles on 10/21/22.
+//
+
+#ifndef VANGUARD_NEARFACTORY_H
+#define VANGUARD_NEARFACTORY_H
+
+
+#include "domain/libBlockchain/langs/LanguageFactory.h"
+
+namespace vanguard {
+
+    class NearFactory : public LanguageFactory {
+    public:
+
+        inline explicit NearFactory(BlockchainDomain::Factory &factory) : LanguageFactory(factory) { }
+
+        BlockchainDomain::Type *createBasicType(const llvm::Module &module, BlockchainModel<BlockchainDomain> &model, std::string name) override;
+
+        Visibility toVisibility(const std::string& visStr) override;
+
+        Mutability toMutability(const std::string& mutStr) override;
+
+    };
+
+}
+
+
+#endif //VANGUARD_NEARFACTORY_H
