+++ lib/domain/libBlockchain/langs/rust/near/NearModel.h
@@ -5,11 +5,12 @@
 #ifndef VANGUARD_NEARMODEL_H
 #define VANGUARD_NEARMODEL_H
 
-#include <analysis/ir-utils/LLVMUtils.h>
-#include <analysis/interproc-analysis/InterProceduralAnalyses.h>
-#include "../../BlockchainModel.h"
+#include "analysis/ir-utils/LLVMUtils.h"
+#include "analysis/interproc-analysis/InterProceduralAnalyses.h"
+#include "domain/libBlockchain/BlockchainModel.h"
 #include "program/TypeClass.h"
 
+// TODO: Remove Domain template parameter
 namespace vanguard {
 
     template <typename Domain>
