+++ lib/domain/libBlockchain/langs/solidity/CMakeLists.txt
@@ -1,2 +1,3 @@
 target_sources(libVanguard PRIVATE
+    SolidityFactory.cpp
 )
