--- lib/domain/libBlockchain/solidity/CMakeLists.txt
@@ -1,1 +1,0 @@
-target_sources(libVanguard PRIVATE )
